chore(litmus-portal): Add unified modal component (#1891)
* Making a unified modal component, Unimodal. This unified modal does not need to be styled again when it needs to be used. Only content passed inside the modal needs to be styled. Can pass flags such as hasCloseBtn to automatically add a close button to the top right, isDark to make the modal background dark and textAlign to set alignment. Most modals have been adapted to use the unified modal
