+++ litmus-portal/frontend/src/components/Sections/Settings/UserManagementTab/EditUser/DelUser/styles.ts
@@ -1,18 +1,6 @@
 import { makeStyles, Theme } from '@material-ui/core/styles';
 
 const useStyles = makeStyles((theme: Theme) => ({
-  modal: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  paper: {
-    width: '50.75rem',
-    height: '39.1875rem',
-    backgroundColor: theme.palette.background.paper,
-    outline: 'none',
-    borderRadius: 3,
-  },
   body: {
     display: 'flex',
     flexDirection: 'column',
