+++ litmus-portal/frontend/src/components/Sections/Settings/UserManagementTab/EditUser/ResetModal/styles.ts
@@ -1,19 +1,6 @@
 import { makeStyles, Theme } from '@material-ui/core/styles';
 
 const useStyles = makeStyles((theme: Theme) => ({
-  modal: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  paper: {
-    width: '50.75rem',
-    height: '39.1875rem',
-    backgroundColor: theme.palette.background.paper,
-
-    outline: 'none',
-    borderRadius: '0.1875rem',
-  },
   body: {
     display: 'flex',
     flexDirection: 'column',
