+++ litmus-portal/frontend/src/components/WorkflowStepper/index.tsx
@@ -9,7 +9,6 @@ import Typography from '@material-ui/core/Typography';
 import { useSelector } from 'react-redux';
 import ButtonFilled from '../Button/ButtonFilled';
 import ButtonOutline from '../Button/ButtonOutline';
-import FinishModal from '../Modal/FinishModal';
 import ReliablityScore from '../Sections/Workflow/ReliabilityScore';
 import ScheduleWorkflow from '../Sections/Workflow/ScheduleWorkflow';
 import VerifyCommit from '../Sections/Workflow/VerifyCommit';
@@ -24,6 +23,7 @@ import { RootState } from '../../redux/reducers';
 import useActions from '../../redux/actions';
 import * as WorkflowActions from '../../redux/actions/workflow';
 import parsed from '../../utils/yamlUtils';
+import Unimodal from '../../containers/layouts/Unimodal';
 
 function getSteps(): string[] {
   return [
@@ -104,6 +104,16 @@ function getStepContent(
 const WorkflowStepper = () => {
   const classes = useStyles();
   const [activeStep, setActiveStep] = React.useState(0);
+  // modal state and handlers
+  const [open, setOpen] = React.useState(false);
+
+  const handleOpen = () => {
+    setOpen(true);
+  };
+
+  const handleClose = () => {
+    setOpen(false);
+  };
 
   const workflowData: WorkflowData = useSelector(
     (state: RootState) => state.workflowData
@@ -176,7 +186,50 @@ const WorkflowStepper = () => {
       <div>
         {activeStep === steps.length ? (
           <div>
-            <FinishModal />
+            <div>
+              <Button
+                variant="contained"
+                color="primary"
+                data-cy="selectModal"
+                onClick={handleOpen}
+              >
+                Completed
+              </Button>
+              {/* FinishModal added */}
+              <Unimodal
+                isOpen={open}
+                handleClose={handleClose}
+                hasCloseBtn={false}
+              >
+                <div className={classes.rootContainer}>
+                  <img
+                    src="icons/finish.png"
+                    className={classes.mark}
+                    alt="mark"
+                  />
+                  <div className={classes.heading}>
+                    A new chaos workflow,
+                    <br />
+                    <strong>was successfully created!</strong>
+                  </div>
+                  <div className={classes.headWorkflow}>
+                    Congratulations on creating your first workflow! Now
+                    information about <br /> it will be displayed on the main
+                    screen of the application.
+                  </div>
+                  <div className={classes.button}>
+                    <Button
+                      variant="contained"
+                      color="secondary"
+                      data-cy="selectFinish"
+                      onClick={handleClose}
+                    >
+                      Back to workflow
+                    </Button>
+                  </div>
+                </div>
+              </Unimodal>
+            </div>
             <Typography className={classes.content}>
               All steps completed (display workflow completed modal here)
             </Typography>
